requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.8
psycopg2-binary==2.9.9
dj-database-url==2.2.0
python-decouple==3.8
anthropic==0.34.2
twilio==9.3.3
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
